Move github local repository logic to local repository access project
