+++ Sources/Application/Kysect.Zeya.IntegrationManager/IGithubRepositoryProvider.cs
@@ -1,5 +1,5 @@
-﻿using Kysect.Zeya.GithubIntegration.Abstraction;
-using Kysect.Zeya.LocalRepositoryAccess;
+﻿using Kysect.Zeya.LocalRepositoryAccess;
+using Kysect.Zeya.LocalRepositoryAccess.Github;
 
 namespace Kysect.Zeya.IntegrationManager;
