+++ Sources/Application/Kysect.Zeya.IntegrationManager/Kysect.Zeya.IntegrationManager.csproj
@@ -12,6 +12,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Kysect.Zeya.GithubIntegration.Abstraction\Kysect.Zeya.GithubIntegration.Abstraction.csproj" />
+    <ProjectReference Include="..\Kysect.Zeya.GitIntegration\Kysect.Zeya.GitIntegration.csproj" />
     <ProjectReference Include="..\Kysect.Zeya.RepositoryValidation.Abstractions\Kysect.Zeya.RepositoryValidation.Abstractions.csproj" />
     <ProjectReference Include="..\Kysect.Zeya.RepositoryValidation\Kysect.Zeya.RepositoryValidation.csproj" />
     <ProjectReference Include="..\Kysect.Zeya.ValidationRules.Abstractions\Kysect.Zeya.ValidationRules.Abstractions.csproj" />
