+++ Sources/Domain/Kysect.Zeya.LocalRepositoryAccess/Github/RepositoryBranchProtection.cs
@@ -1,3 +1,3 @@
-﻿namespace Kysect.Zeya.GithubIntegration.Abstraction;
+﻿namespace Kysect.Zeya.LocalRepositoryAccess.Github;
 
 public record struct RepositoryBranchProtection(bool PullRequestReviewsRequired, bool ConversationResolutionRequired);
