+++ Sources/Presentation/Kysect.Zeya.Tui/Commands/AnalyzeAndFixAndCreatePullRequestRepositoryCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.TerminalUserInterface.Commands;
-using Kysect.Zeya.GithubIntegration.Abstraction;
 using Kysect.Zeya.IntegrationManager;
+using Kysect.Zeya.LocalRepositoryAccess.Github;
 using Kysect.Zeya.Tui.Controls;
 
 namespace Kysect.Zeya.Tui.Commands;
