+++ Sources/Tests/Kysect.Zeya.Tests/Tools/Fakes/FakeGithubIntegrationService.cs
@@ -4,6 +4,7 @@ using Kysect.GithubUtils.Replication.RepositorySync.LocalStoragePathFactories;
 using Kysect.Zeya.GithubIntegration;
 using Kysect.Zeya.GithubIntegration.Abstraction;
 using Kysect.Zeya.LocalRepositoryAccess;
+using Kysect.Zeya.LocalRepositoryAccess.Github;
 using Microsoft.Extensions.Logging;
 
 namespace Kysect.Zeya.Tests.Tools.Fakes;
