+++ Sources/Tests/Kysect.Zeya.Tests/ValidationRules/Github/GithubBranchProtectionEnabledValidationRuleTests.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.ScenarioLib.Abstractions;
-using Kysect.Zeya.GithubIntegration.Abstraction;
 using Kysect.Zeya.LocalRepositoryAccess;
+using Kysect.Zeya.LocalRepositoryAccess.Github;
 using Kysect.Zeya.RepositoryValidation.Abstractions;
 using Kysect.Zeya.Tests.Tools;
 using Kysect.Zeya.Tests.Tools.Fakes;
